requirements.txt
django>=4.2,<5.0
djangorestframework




markdown
django-filter
drf-yasg[validation]
djangorestframework-simplejwt


black
flake8